+++ doc/news.rst
@@ -26,6 +26,8 @@ you should review this document and see if your consfig needs updating.
 1.3.2 (unreleased)
 ------------------
 
+- Add LIBVIRT:KVM-BOOTS-LVM-LV and LIBVIRT:KVM-BOOTS-LVM-LV-FOR.
+
 - Fix a bug in DISK:HOST-LOGICAL-VOLUMES-EXIST that meant it would try to add
   ``/etc/fstab`` entries for each logical volume it created rather than for
   each filesystem it tries to mount.